Assignment 1 | data type changed to long and formatted the code

+++ src/main/java/com/zs/assignments/assignment1/SystemDetails.java
@@ -3,36 +3,42 @@ package com.zs.assignments.assignment1;
 import java.io.File;
 
 import java.lang.management.ManagementFactory;
+
 import com.sun.management.OperatingSystemMXBean;
 
 public class SystemDetails {
     OperatingSystemMXBean osBean = (OperatingSystemMXBean) ManagementFactory.getOperatingSystemMXBean();
     File root = new File("/");
 
-    public String getUser(){
+    public String getUser() {
         return System.getProperty("user.name");
     }
-    public String getHomeDirectory(){
+
+    public String getHomeDirectory() {
         return System.getProperty("user.home");
     }
-    public String getOsVersion(){
+
+    public String getOsVersion() {
         return System.getProperty("os.version");
     }
-    public String getOsBuild(){
+
+    public String getOsBuild() {
         return System.getProperty("os.name");
     }
 
-    public int getSystemMemory(){
+    public long getSystemMemory() {
         long totalMemoryInBytes = osBean.getTotalPhysicalMemorySize();
-        int totalMemoryInGB = (int) (totalMemoryInBytes/(1024*1024*1024));
+        long totalMemoryInGB = (totalMemoryInBytes / (1024 * 1024 * 1024));
         return totalMemoryInGB;
     }
-    public int getSystemDiskSize(){
+
+    public long getSystemDiskSize() {
         long totalDiskSpaceInBytes = root.getTotalSpace();
-        int totalDiskSpaceInGB = (int)(totalDiskSpaceInBytes/(1024*1024*1024));
+        long totalDiskSpaceInGB = (int) (totalDiskSpaceInBytes / (1024 * 1024 * 1024));
         return totalDiskSpaceInGB;
     }
-    public int getSystemCores(){
+
+    public int getSystemCores() {
         return osBean.getAvailableProcessors();
     }
 }
